pyproject.toml
[tool.poetry]
name = "trktor"
version = "0.1.0"
description = ""
authors = ["nomachi <nomachi@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "==2.5.0"
pillow = "==11.1.0"
pendulum = "==3.0.0"
pydantic = "==2.11.0a2"
click = "==8.1.8"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "==1.15.0"
ruff = "==0.9.6"
python-semantic-release = "==9.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]   # Add flake8-bugbear rules
ignore = ["E501"]                       # Ignore line-length violations
